requirements.txt
requests
pandas
jinja2
python-slugify
